ApproxFun.jl Documentation

Domains

ApproxFun.Arc — Type.

Arc(c,r,(θ₁,θ₂))

represents the arc centred at c with radius r from angle θ₁ to θ₂.

source

ApproxFun.Circle — Type.

Circle(c,r,o)

represents the circle centred at c with radius r which is positively (o=true) or negatively (o=false) oriented.

source

ApproxFun.Curve — Constant.

Curve Represents a domain defined by the image of a Fun. Example usage would be

x=Fun([1,2])
Curve(exp(im*x))  # represents an arc

source

ApproxFun.Disk — Type.

Disk(c,r)

represents the disk centred at c with radius r.

source

ApproxFun.Interval — Type.

Interval(a,b)

represents an interval from a to b. In the case where a and b are complex or 2-dimensional, it represents the line segment between a and b.

source

ApproxFun.Line — Type.

Line{a}(c)

represents the line at angle a in the complex plane, centred at c.

source

ApproxFun.PeriodicInterval — Type.

PeriodicInterval(a,b)

represents a periodic interval from a to b, that is, the point b is identified with a.

source

ApproxFun.ProductDomain — Type.

ProductDomain((d1,d2))

represents the product of two domains, the set {(x,y) : x ∈ d1 & y ∈ d2}.

Multiplication of domains is overrident to return a ProductDomain. For example, the following represents the rectangle 1 ≤ x ≤ 2 & 3 ≤ y ≤ 4:

Interval(1,2)*(3,4)

source

ApproxFun.Ray — Type.

Ray{a}(c,o)

represents a ray at angle a starting at c, with orientation out to infinity (o = true) or back from infinity (o = false).

source

ApproxFun.UnionDomain — Type.

UnionDomain((d1,d2,…,dn))

represents a union of multiple subdomains: {x : x ∈ d1 || … || x ∈ dn}.

source

ApproxFun.∂ — Function.

∂(domain)

returns the boundary of domain. For example, the boundary of a Disk() is a Circle(), and the boundary of Interval()^2 is a piecewise interval that sketches the boundary of a rectangle.

source

Constructing a Fun

ApproxFun.Fun — Type.

Fun(coefficients,space)

returns a Fun with coefficients in the space

source

Accessing information about a Fun

ApproxFun.domain — Function.

domain(::Fun)

returns the domain that a Fun is defined on

source

ApproxFun.coefficients — Function.

coefficients(fun,space)

returns the coefficients of a fun in a possibly different space

source

ApproxFun.extrapolate — Function.

extrapolate(fun,x)

returns an extrapolation of fun from its domain to x.

source

ApproxFun.ncoefficients — Function.

ncoefficients(fun) -> Integer

returns the number of coefficients of a fun

source

ApproxFun.points — Function.

points(fun)

returns a grid of points that the fun can be transformed into values and back

source

ApproxFun.space — Function.

space(fun)

returns the space of fun

source

Base.values — Function.

values(fun)

returns fun evaluated at points(fun)

source

Modify a Fun

ApproxFun.reverseorientation — Function.

reverseorientation(fun)

return fun on a reversed orientated contour

source

ApproxFun.setdomain — Function.

setdomain(fun,domain)

returns fun projected onto domain

source

Base.chop — Function.

chop(fun,tol)

reduces the number of coefficients by dropping the tail that is below the specified tolerance.

source

Operators

ApproxFun.Operator — Type.

Operator{T}

is an abstract type to represent linear operators between spaces.

source

ApproxFun.linsolve — Function.

linsolve(A,b;tolerance=tol,maxlength=n)

solves a linear equation, usually differential equation, where A is an operator or array of operators and b is a Fun or array of funs. The result u will approximately satisfy A*u = b.

source

Base.LinAlg.qrfact — Method.

qrfact(A::Operator)

returns a cached QR factorization of the Operator A. The result QR enables solving of linear equations: if u=QR\b, then u approximately satisfies A*u = b.

source

ApproxFun.cache — Method.

cache(operator)

Caches the entries of an operator, to speed up multiplying a Fun by the operator.

source