ApproxFun.jl Documentation

Domains

ApproxFun.Interval — Type.

Interval(a,b)

represents an interval from a to b. In the case where a and b are complex or 2-dimensional, it represents the line segment between a and b.

Constructing a Fun

ApproxFun.Fun — Type.

Fun(coefficients,space)

returns a Fun with coefficients in the space

Accessing information about a Fun

ApproxFun.domain — Function.

domain(::Fun)

returns the domain that a Fun is defined on

ApproxFun.coefficients — Function.

coefficients(fun,space)

returns the coefficients of a fun in a possibly different space

ApproxFun.extrapolate — Function.

extrapolate(fun,x)

returns an extrapolation of fun from its domain to x.

ApproxFun.ncoefficients — Function.

ncoefficients(fun) -> Integer

returns the number of coefficients of a fun

ApproxFun.points — Function.

points(fun)

returns a grid of points that the fun can be transformed into values and back

ApproxFun.space — Function.

space(fun)

returns the space of fun

Base.values — Function.

values(fun)

returns fun evaluated at points(fun)

Modify a Fun

ApproxFun.reverseorientation — Function.

reverseorientation(fun)

return fun on a reversed orientated contour

ApproxFun.setdomain — Function.

setdomain(fun,domain)

returns fun projected onto domain

Operators

ApproxFun.Operator — Type.

Operator{T}

is an abstract type to represent linear operators between spaces.

ApproxFun.linsolve — Function.

linsolve(A,b;tolerance=tol,maxlength=n)

solves a linear equation, usually differential equation, where A is an operator or array of operators and b is a Fun or array of funs. The result u will approximately satisfy A*u = b.

Base.LinAlg.qrfact — Method.

qrfact(A::Operator)

returns a cached QR factorization of the Operator A. The result QR enables solving of linear equations: if u=QR\b, then u approximately satisfies A*u = b.