Library

Library

Constructing a Fun

ApproxFun.Fun — Type.

Fun(s::Space,coefficients::AbstractVector)

returns a Fun with the specified coefficients in the space s

source

Fun(f,s::Space)

return a Fun representing the function, number, or vector f in the space s. If f is vector-valued, it returns a vector-valued analogue of s.

source

Fun(f,d::Domain)

returns Fun(f,Space(d)), that is, it uses the default space for the specified domain.

source

Fun(s::Space)

returns Fun(identity,s)

source

Fun(f)

returns Fun(f,Chebyshev())

source

Fun()

returns Fun(identity,Chebyshev()).

source

Base.ones — Method.

ones(d::Space)

Return the Fun that represents the function one on the specified space.

source

Base.zeros — Method.

zeros(d::Space)

Return the Fun that represents the function one on the specified space.

source

Domains

ApproxFun.Arc — Type.

Arc(c,r,(θ₁,θ₂))

represents the arc centred at c with radius r from angle θ₁ to θ₂.

source

ApproxFun.Circle — Type.

Circle(c,r,o)

represents the circle centred at c with radius r which is positively (o=true) or negatively (o=false) oriented.

source

ApproxFun.Curve — Constant.

Curve Represents a domain defined by the image of a Fun. Example usage would be

x=Fun(1..2)
Curve(exp(im*x))  # represents an arc

source

ApproxFun.Disk — Type.

Disk(c,r)

represents the disk centred at c with radius r.

source

ApproxFun.Segment — Type.

Segment(a,b)

represents a line segment from a to b. In the case where a and b are real and a < b, then this is is equivalent to an Interval(a,b).

source

ApproxFun.Interval — Function.

Interval(a::Real,b::Real)

represents the set {x : a ≤ x ≤ b}.

source

ApproxFun.Line — Type.

Line{a}(c)

represents the line at angle a in the complex plane, centred at c.

source

ApproxFun.PeriodicInterval — Type.

PeriodicInterval(a,b)

represents a periodic interval from a to b, that is, the point b is identified with a.

source

ApproxFun.Point — Type.

Point(x)

represents a single point at x.

source

ApproxFun.ProductDomain — Type.

ProductDomain((d1,d2))

represents the product of two domains, the set {(x,y) : x ∈ d1 & y ∈ d2}.

Multiplication of domains is overrident to return a ProductDomain. For example, the following represents the rectangle 1 ≤ x ≤ 2 & 3 ≤ y ≤ 4:

Interval(1,2)*(3,4)

source

ApproxFun.Ray — Type.

Ray{a}(c,o)

represents a ray at angle a starting at c, with orientation out to infinity (o = true) or back from infinity (o = false).

source

ApproxFun.UnionDomain — Type.

UnionDomain((d1,d2,…,dn))

represents a union of multiple subdomains: {x : x ∈ d1 || … || x ∈ dn}.

source

ApproxFun.∂ — Function.

∂(d::Domain)

returns the boundary of d. For example, the boundary of a Disk() is a Circle(), and the boundary of Interval()^2 is a piecewise interval that sketches the boundary of a rectangle.

source

Accessing information about a spaces

ApproxFun.canonicalspace — Function.

canonicalspace(s::Space)

returns a space that is used as a default to implement missing functionality, e.g., evaluation. Implement a Conversion operator or override coefficients to support this.

source

ApproxFun.itransform — Function.

itransform(s::Space,coefficients::AbstractVector)

Transform coefficients back to values. Defaults to using canonicalspace as in transform.

source

ApproxFun.transform — Function.

transform(s::Space,vals::Vector)

Transform values on the grid specified by points(s,length(vals)) to coefficients in the space s. Defaults to coefficients(transform(canonicalspace(space),values),canonicalspace(space),space)

source

ApproxFun.evaluate — Method.

evaluate(sp::Space,coefficients::AbstractVector,x)

Evaluates the expansion at a point x. If x is in the domain, then this should return zero.

source

ApproxFun.dimension — Method.

dimension(s::Space)

returns the dimension of s, which is the maximum number of coefficients.

source

Inbuilt spaces

ApproxFun.SequenceSpace — Type.

SequenceSpace is the space of all sequences, i.e., infinite vectors. Also denoted ℓ⁰.

source

ApproxFun.ConstantSpace — Type.

ConstantSpace is the 1-dimensional scalar space.

source

ApproxFun.Chebyshev — Type.

Chebyshev() is the space spanned by the Chebyshev polynomials

    T_0(x),T_1(x),T_2(x),…

where T_k(x) = cos(k*acos(x)). This is the default space as there exists a fast transform and general smooth functions on [-1,1] can be easily resolved.

source

ApproxFun.Jacobi — Type.

Jacobi(b,a) represents the space spanned by Jacobi polynomials P_k^{(a,b)}, which are orthogonal with respect to the weight (1+x)^β*(1-x)^α

source

ApproxFun.Ultraspherical — Type.

Ultraspherical(λ) is the space spanned by the ultraspherical polynomials

    C_0^{(λ)}(x),C_1^{(λ)}(x),C_2^{(λ)}(x),…

Note that λ=1 this reduces to Chebyshev polynomials of the second kind: C_k^{(1)}(x) = U_k(x).

source

ApproxFun.Taylor — Type.

Taylor() is the space spanned by [1,z,z^2,...]. This is a type alias for Hardy{true}.

source

ApproxFun.Hardy — Type.

Hardy{false}() is the space spanned by [1/z,1/z^2,...]. Hardy{true}() is the space spanned by [1,z,z^2,...].

source

ApproxFun.Fourier — Type.

Fourier() is the space spanned by the trigonemtric polynomials

    1,sin(θ),cos(θ),sin(2θ),cos(2θ),…

See also Laurent.

source

ApproxFun.Laurent — Type.

Laurent() is the space spanned by the complex exponentials

    1,exp(-im*θ),exp(im*θ),exp(-2im*θ),…

See also Fourier.

source

ApproxFun.CosSpace — Type.

CosSpace() is the space spanned by [1,cos θ,cos 2θ,...]

source

ApproxFun.SinSpace — Type.

SinSpace() is the space spanned by [sin θ,sin 2θ,...]

source

ApproxFun.JacobiWeight — Type.

JacobiWeight(β,α,s::Space)

weights a space s by a Jacobi weight, which on -1..1 is (1+x)^β*(1-x)^α. For other domains, the weight is inferred by mapping to -1..1.

source

ApproxFun.LogWeight — Type.

LogWeight(β,α,s::Space)

represents a function on -1..1 weighted by log((1+x)^β*(1-x)^α). For other domains, the weight is inferred by mapping to -1..1.

source

ApproxFun.ArraySpace — Type.

ArraySpace(s::Space,dims...)

is used to represent array-valued expansions in a space s. The coefficients are of each entry are interlaced.

For example,

f = Fun(x->[exp(x),sin(x)],-1..1)
space(f) == ArraySpace(Chebyshev(),2)

source

ApproxFun.TensorSpace — Type.

TensorSpace(a::Space,b::Space)

represents a tensor product of two 1D spaces a and b. The coefficients are interlaced in lexigraphical order.

For example, consider

Fourier()*Chebyshev()  # returns TensorSpace(Fourier(),Chebyshev())

This represents functions on [-π,π) x [-1,1], using the Fourier basis for the first argument and Chebyshev basis for the second argument, that is, φ_k(x)T_j(y), where

φ_0(x) = 1,
φ_1(x) = sin x,
φ_2(x) = cos x,
φ_3(x) = sin 2x,
φ_4(x) = cos 2x
…

By Choosing (k,j) appropriately, we obtain a single basis:

φ_0(x)T_0(y) (= 1),
φ_0(x)T_1(y) (= y),
φ_1(x)T_0(y) (= sin x),
φ_0(x)T_2(y), …

source

Accessing information about a Fun

ApproxFun.domain — Function.

domain(f::Fun)

returns the domain that f is defined on

source

ApproxFun.coefficients — Function.

coefficients(f::Fun) -> Vector

returns the coefficients of f, corresponding to the space space(f).

source

coefficients(f::Fun,s::Space) -> Vector

returns the coefficients of f in the space s, which may not be the same as space(f).

source

coefficients(cfs::AbstractVector,fromspace::Space,tospace::Space) -> Vector

converts coefficients in fromspace to coefficients in tospace

source

ApproxFun.extrapolate — Function.

extrapolate(f::Fun,x)

returns an extrapolation of f from its domain to x.

source

ApproxFun.ncoefficients — Function.

ncoefficients(f::Fun) -> Integer

returns the number of coefficients of a fun

source

ApproxFun.points — Function.

points(f::Fun)

returns a grid of points that f can be transformed into values and back

source

points(s::Space,n::Integer)

returns a grid of approximately n points, for which a transform exists from values at the grid to coefficients in the space s.

source

ApproxFun.space — Function.

space(f::Fun)

returns the space of f

source

Base.values — Function.

values(f::Fun)

returns f evaluated at points(f)

source

Base.stride — Method.

stride(f::Fun)

returns the stride of the coefficients, checked numerically

source

Modify a Fun

ApproxFun.reverseorientation — Function.

reverseorientation(f::Fun)

return f on a reversed orientated contour.

source

ApproxFun.setdomain — Function.

setdomain(f::Fun,d::Domain)

returns f projected onto domain

source

Base.chop — Function.

chop(f::Fun,tol) -> Fun

reduces the number of coefficients by dropping the tail that is below the specified tolerance.

source

Operators

ApproxFun.Operator — Type.

Operator{T}

is an abstract type to represent linear operators between spaces.

source

BandedMatrices.bandwidths — Method.

bandwidths(op::Operator)

returns the bandwidth of op in the form (l,u), where l ≥ 0 represents the number of subdiagonals and u ≥ 0 represents the number of superdiagonals.

source

ApproxFun.domainspace — Function.

domainspace(op::Operator)

gives the domain space of op. That is, op*f will first convert f to a Fun in the space domainspace(op) before applying the operator.

source

ApproxFun.rangespace — Function.

rangespace(op::Operator)

gives the range space of op. That is, op*f will return a Fun in the space rangespace(op), provided f can be converted to a Fun in domainspace(op).

source

Base.getindex — Method.

op[k,j]

returns the kth coefficient of op*Fun([zeros(j-1);1],domainspace(op)).

source

Base.getindex — Method.

op[f::Fun]

constructs the operator op*Multiplication(f), that is, it multiplies on the right by f first. Note that op*f is different: it applies op to f.

source

Base.:\ — Method.

\(A::Operator,b;tolerance=tol,maxlength=n)

solves a linear equation, usually differential equation, where A is an operator or array of operators and b is a Fun or array of funs. The result u will approximately satisfy A*u = b.

source

Base.LinAlg.qrfact — Method.

qrfact(A::Operator)

returns a cached QR factorization of the Operator A. The result QR enables solving of linear equations: if u=QR\b, then u approximately satisfies A*u = b.

source

ApproxFun.cache — Method.

cache(op::Operator)

Caches the entries of an operator, to speed up multiplying a Fun by the operator.

source

Inbuilt operators

ApproxFun.Conversion — Type.

Conversion(fromspace::Space,tospace::Space)

represents a conversion operator between fromspace and tospace, when available.

source

ApproxFun.Derivative — Type.

Derivative(sp::Space,k::Int) represents the k-th derivative on sp.

source

Derivative(sp::Space,k::Vector{Int}) represents a partial derivative on a multivariate space. For example,

Dx = Derivative(Chebyshev()^2,[1,0]) # ∂/∂x
Dy = Derivative(Chebyshev()^2,[0,1]) # ∂/∂y

source

Derivative(sp::Space) represents the first derivative Derivative(sp,1).

source

Derivative(k) represents the k-th derivative, acting on an unset space. Spaces will be inferred when applying or manipulating the operator.

source

Derivative() represents the first derivative on an unset space. Spaces will be inferred when applying or manipulating the operator.

source

ApproxFun.Dirichlet — Type.

Dirichlet(sp,k) is the operator associated with restricting the k-th derivative on the boundary for the space sp.

source

Dirichlet(sp) is the operator associated with restricting the the boundary for the space sp.

source

Dirichlet() is the operator associated with restricting on the the boundary.

source

ApproxFun.Evaluation — Type.

Evaluation(sp,x,k) is the functional associated with evaluating the k-th derivative at a point x for the space sp.

source

Evaluation(sp,x) is the functional associated with evaluating at a point x for the space sp.

source

Evaluation(x) is the functional associated with evaluating at a point x.

source

ApproxFun.Integral — Type.

Integral(sp::Space,k::Int) represents a k-th integral on sp. There is no guarantee on normalization.

source

Integral(sp::Space) represents the first integral Integral(sp,1).

source

Integral(k)represents thek`-th integral, acting on an unset space. Spaces will be inferred when applying or manipulating the operator.

source

Intergral() represents the first integral on an unset space. Spaces will be inferred when applying or manipulating the operator.

source

ApproxFun.Laplacian — Type.

Laplacian(sp::Space) represents the laplacian on space sp.

source

Laplacian() represents the laplacian on an unset space. Spaces will be inferred when applying or manipulating the operator.

source

ApproxFun.Multiplication — Type.

Multiplication(f::Fun,sp::Space) is the operator representing multiplication by f on functions in the space sp.

source

Multiplication(f::Fun) is the operator representing multiplication by f on an unset space of functions. Spaces will be inferred when applying or manipulating the operator.

source

ApproxFun.Neumann — Function.

Neumann(sp) is the operator associated with restricting the normal derivative on the boundary for the space sp. At the moment it is implemented as Dirichlet(sp,1).

source

`Neumann( is the operator associated with restricting the normal derivative on the boundary.

source

Bivariate

ApproxFun.LowRankFun — Type.

LowRankFun gives an approximation to a bivariate function in low rank form.

source